i18n/requirements.txt
deepl==1.22.0
python-dotenv==1.0.0
argparse 